% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_query.R, R/explore_dev.R
\name{screenshot}
\alias{screenshot}
\title{screenshot}
\usage{
screenshot(browser, display = T, useViewer = T, file = NULL)

screenshot(browser, display = T, useViewer = T, file = NULL)
}
\description{
screenshot

screenshot
}
